manthana/requirements.txt
asgiref==3.5.0
astroid==2.9.3
backports.zoneinfo==0.2.1
cairocffi==1.2.0
CairoSVG==2.5.2
captcha==0.4
certifi==2020.6.20
cffi==1.14.6
chardet==3.0.4
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.12
django-captcha==1.0.0
django-dirtyfields==1.8.1
django-jinja==2.10.0
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-report-builder==6.4.2
django-simple-captcha==0.5.17
djangorestframework==3.12.4
et-xmlfile==1.1.0
html5lib==1.1
idna==2.10
isort==5.10.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.0
mccabe==0.6.1
mysql-client==0.0.1
mysql-connector-python==8.0.21
#mysqlclient==2.0.1
openpyxl==3.0.7
Pillow==8.3.1
platformdirs==2.5.1
protobuf==3.12.4
pycparser==2.20
pylint==2.12.2
pyphen==0.11.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tinycss2==1.1.0
toml==0.10.2
typing-extensions==4.1.1
urllib3==1.25.10
WeasyPrint==52.5
webencodings==0.5.1
wrapt==1.13.3
xlwt==1.3.0
